+++ selene/support/past/__init__.py
@@ -25,11 +25,4 @@ from selene.support.shared import browser, config
 from selene.support.past.support import by  # todo: make analog in selene.*
 from selene.support.conditions import be, have
 from selene.condition import not_
-
-
-def s(css_selector_or_by: str) -> Element:
-    return browser.element(css_selector_or_by)
-
-
-def ss(css_selector_or_by: str) -> Collection:
-    return browser.all(css_selector_or_by)
+from selene.support.jquery_style_selectors import s, ss