+++ tests/examples/todomvc/pagemodules_approach/pages/tasks.py
@@ -2,7 +2,7 @@ from selene.support.past.bys import by_link_text
 from selene.support.past.conditions import exact_text, visible, exact_texts, enabled
 from selene.support.conditions import have
 from selene.support.past import browser
-from selene.support.past.support.jquery_style_selectors import s, ss
+from selene.support.jquery_style_selectors import s, ss
 
 __author__ = 'yashaka'