+++ tests/integration/error_messages_test.py
@@ -29,7 +29,7 @@ from webdriver_manager.chrome import ChromeDriverManager
 from selene.support.past import browser, config
 from selene.support.past.common.none_object import NoneObject
 from selene.support.conditions import have
-from selene.support.past.support.jquery_style_selectors import s, ss
+from selene.support.jquery_style_selectors import s, ss
 from tests.integration.helpers.givenpage import GivenPage
 
 GIVEN_PAGE = NoneObject('GivenPage')  # type: GivenPage