+++ tests/integration/test_find_from_hidden_element.py
@@ -25,7 +25,7 @@ import os
 from selene.support.past import browser, config
 from selene.support.past.conditions import in_dom, hidden, text, size
 from selene.support.past.support import by
-from selene.support.past.support.jquery_style_selectors import s
+from selene.support.jquery_style_selectors import s
 
 start_page = 'file://' + os.path.abspath(os.path.dirname(__file__)) + '/../resources/start_page.html'