+++ tests/integration/test_unicode.py
@@ -28,7 +28,7 @@ from selene.support.past import browser, config
 from selene.support.past.browsers import BrowserName
 from selene.support.past.conditions import texts, exact_text
 from selene.support.conditions import have
-from selene.support.past.support.jquery_style_selectors import s, ss
+from selene.support.jquery_style_selectors import s, ss
 
 start_page = 'file://' + os.path.abspath(os.path.dirname(__file__)) + '/../resources/start_page.html'