+++ tests/past/acceptance/helpers/todomvc.py
@@ -27,7 +27,7 @@ from selenium.webdriver.support.wait import WebDriverWait
 from selene.support.past import config
 from selene.support.past.conditions import visible
 from selene.support.past.browser import driver
-from selene.support.past.support.jquery_style_selectors import s
+from selene.support.jquery_style_selectors import s
 from selene.support.past.wait import satisfied
 
 __author__ = 'yashaka'