+++ tests/past/acceptance/mixed_driver_management/f_automatic_shared_driver_test.py
@@ -23,7 +23,7 @@
 from selene.support.past import browser
 from selene.support.conditions import have
 
-from selene.support.past.support.jquery_style_selectors import s, ss
+from selene.support.jquery_style_selectors import s, ss
 
 
 todomvc_url = 'https://todomvc4tasj.herokuapp.com/'