+++ tests/past/acceptance/selene_element_test.py
@@ -25,7 +25,7 @@ from selenium.common.exceptions import TimeoutException
 
 from selene.support.past import config
 from selene.support.past.conditions import hidden, exist
-from selene.support.past.support.jquery_style_selectors import s
+from selene.support.jquery_style_selectors import s
 from tests.past.acceptance.helpers.helper import get_test_driver
 
 __author__ = 'yashaka'