+++ tests/past/acceptance/selene_page_factory.py
@@ -26,7 +26,7 @@ import selene
 from selene.support.past.browsers import BrowserName
 from selene.support.past.conditions import exact_text
 from selene.support.past.browser import open_url
-from selene.support.past.support.jquery_style_selectors import s
+from selene.support.jquery_style_selectors import s
 
 start_page = 'file://' + os.path.abspath(os.path.dirname(__file__)) + '/../resources/start_page.html'