+++ tests/past/acceptance/speed_comparison_of_simple_element_actions_test.py
@@ -27,7 +27,7 @@ from selenium.webdriver.support.wait import WebDriverWait
 
 from selene.support.past.conditions import visible
 from selene.support.past.browser import set_driver, driver, open_url
-from selene.support.past.support.jquery_style_selectors import s
+from selene.support.jquery_style_selectors import s
 from tests.past.acceptance.helpers.helper import get_test_driver
 from tests.helpers import time_spent
 from tests.past.acceptance.helpers.todomvc import TODOMVC_URL