+++ CHANGELOG.md
@@ -7,8 +7,10 @@
     - reflected all "old and redundant" SeleneCollection methods as deprecated in Collection
       - did not add methods that were already deprecated in 1.*
     - temporally added selene.browser module to reflect browser.* methods from 1.* as deprecated
+    - moved jquery_style_selectors.py module back to selene.support
 - fixed autocomplete for *.should methods
 - fixed browser.switch_to
+
 
 ## 2.0.0a4 (released on 30.12.2019)
 - fixed default browser_name handling in shared config; implemented some old opts in config